Rename backend modules and remove extra files
