--- shopping/.idea/inspectionProfiles/Project_Default.xml
@@ -1,15 +1,0 @@
-<component name="InspectionProjectProfileManager">
-  <profile version="1.0">
-    <option name="myName" value="Project Default" />
-    <inspection_tool class="PyUnresolvedReferencesInspection" enabled="true" level="WARNING" enabled_by_default="true">
-      <option name="ignoredIdentifiers">
-        <list>
-          <option value="list.get_attribute" />
-          <option value="fullstack_django.fullstack_django.urls.backend_api" />
-          <option value="fullstack_django.backend_api.views.product_list1" />
-          <option value="main" />
-        </list>
-      </option>
-    </inspection_tool>
-  </profile>
-</component>
